# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CategoryOrder:
    properties:
      id:
        type: string
      sort_order:
        type: integer
    required:
    - id
    - sort_order
    type: object
  dto.CreateCategoryRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 120
        type: string
      sort_order:
        type: integer
    required:
    - name
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  dto.LogoutRequest:
    properties:
      refreshToken:
        type: string
    type: object
  dto.RefreshTokenRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  dto.RegisterRequest:
    properties:
      email:
        type: string
      fullName:
        maxLength: 150
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - fullName
    - password
    type: object
  dto.ReorderCategoryRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.CategoryOrder'
        minItems: 1
        type: array
    required:
    - items
    type: object
  dto.ToggleCategoryActiveRequest:
    properties:
      is_active:
        type: boolean
    required:
    - is_active
    type: object
...
  dto.UpdateCategoryRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 120
        type: string
    type: object
...
  response.Envelope:
    properties:
      data: {}
      error: {}
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: Backend API Service untuk platform SaaS POS & Self-Order UMKM kuliner.
  title: saas_gangsta Backend API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login untuk semua role (CUSTOMER, PARTNER, ADMIN)
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Envelope'
      summary: Login user
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user dari sesi aktif
      parameters:
      - description: Logout payload
        in: body
        name: request
        schema:
          $ref: '#/definitions/dto.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - Auth
  /auth/me:
    get:
      description: Ambil profil user yang sedang login
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Current user info
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh access token menggunakan refresh token yang valid
      parameters:
      - description: Refresh payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Envelope'
      summary: Refresh token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Membuat akun baru untuk login dengan role default customer
      parameters:
      - description: Register payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Envelope'
      summary: Register user account
      tags:
      - Auth
  /categories:
    get:
      description: Mengambil daftar category per tenant
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List Categories
      tags:
      - Category Management
    post:
      consumes:
      - application/json
      description: Buat category baru
      parameters:
      - description: Payload Create
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create Category
      tags:
      - Category Management
  /categories/{id}:
    delete:
      description: Hapus category (soft delete)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Soft Delete Category
      tags:
      - Category Management
    get:
      description: Mengambil detail satu category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Detail Category
      tags:
      - Category Management
    put:
      consumes:
      - application/json
      description: Perbarui nama atau deskripsi category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Payload Update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update Category
      tags:
      - Category Management
  /categories/{id}/toggle-active:
    patch:
      consumes:
      - application/json
      description: Ubah status is_active category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Payload Toggle
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ToggleCategoryActiveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Aktif / Nonaktifkan Category
      tags:
      - Category Management
  /categories/reorder:
    patch:
      consumes:
      - application/json
      description: Ubah urutan (sort_order) untuk beberapa category sekaligus
      parameters:
      - description: Payload Reorder
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ReorderCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update sort_order
      tags:
      - Category Management
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"